requirements.txt
aiosmtplib==2.0.1
anyio==3.6.2
bcrypt==3.2.0
blinker==1.6.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
cryptography==40.0.2
dnspython==2.3.0
ecdsa==0.18.0
email-validator==1.3.1
fastapi==0.95.1
fastapi-mail==1.2.7
future==0.15.2
greenlet==1.1.2
h11==0.14.0
httpcore==0.17.0
httptools==0.5.0
httpx==0.24.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
mangum==0.17.0
MarkupSafe==2.1.2
openpay==1.1
orjson==3.8.10
passlib==1.7.4
psycopg2-binary==2.9.5
pyasn1==0.5.0
pycparser==2.21
pydantic==1.10.4
PyJWT==2.6.0
PyMySQL==1.0.2
python-dotenv==0.21.1
python-jose==3.3.0
python-multipart==0.0.6
PyYAML==6.0
requests==2.28.2
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.46
starlette==0.26.1
typing_extensions==4.4.0
ujson==5.7.0
urllib3==1.26.14
uvicorn==0.20.0
uvloop==0.17.0
watchfiles==0.19.0
websockets==11.0.2
